+++ rapyer/types/datetime.py
@@ -3,8 +3,9 @@ from typing import TYPE_CHECKING
 
 from pydantic_core import core_schema
 from pydantic_core.core_schema import ValidationInfo, SerializationInfo
-from rapyer.types.base import RedisType, REDIS_DUMP_FLAG_NAME
 from redis.commands.search.field import NumericField
+
+from rapyer.types.base import RedisType, REDIS_DUMP_FLAG_NAME
 
 
 class RedisDatetime(datetime, RedisType):
@@ -71,5 +72,5 @@ class RedisDatetimeTimestamp(RedisDatetime):
 
 
 if TYPE_CHECKING:
-    RedisDatetime = RedisDatetime | datetime
-    RedisDatetimeTimestamp = RedisDatetimeTimestamp | datetime
+    RedisDatetime = RedisDatetime | datetime  # pragma: no cover
+    RedisDatetimeTimestamp = RedisDatetimeTimestamp | datetime  # pragma: no cover