+++ rapyer/types/float.py
@@ -1,7 +1,8 @@
 from typing import TypeAlias, TYPE_CHECKING
 
-from rapyer.types.base import RedisType
 from redis.commands.search.field import NumericField
+
+from rapyer.types.base import RedisType
 
 
 class RedisFloat(float, RedisType):
@@ -44,4 +45,4 @@ class RedisFloat(float, RedisType):
 
 
 if TYPE_CHECKING:
-    RedisFloat: TypeAlias = RedisFloat | float
+    RedisFloat: TypeAlias = RedisFloat | float  # pragma: no cover