+++ rapyer/types/integer.py
@@ -1,8 +1,9 @@
 from typing import TypeAlias, TYPE_CHECKING
 
-from rapyer.types.base import RedisType
 from redis.commands.search.field import NumericField
 from typing_extensions import deprecated
+
+from rapyer.types.base import RedisType
 
 
 class RedisInt(int, RedisType):
@@ -63,4 +64,4 @@ class RedisInt(int, RedisType):
 
 
 if TYPE_CHECKING:
-    RedisInt: TypeAlias = RedisInt | int
+    RedisInt: TypeAlias = RedisInt | int  # pragma: no cover